+++ plottr/apps/inspectr.py
@@ -444,7 +444,7 @@ class QCodesDBInspector(QtWidgets.QMainWindow):
     def setMonitorInterval(self, val: float) -> None:
         self.monitor.stop()
         if val > 0:
-            self.monitor.start(val * 1000)
+            self.monitor.start(int(val * 1000))
 
         self.monitorInput.spin.setValue(val)
